Revert "remove events outside duration before dedupe and (especially) compression, as we don't want to compress events that shouldn't be included."
This reverts commit a4b307ebd61fcc9d80b1ecb1d0177312a758c0da.
Reason is, under further consideration, it makes more sense to compress all events first, regardless if the event time is outside the duration.
The idea here is that at the earliest event within the duration, you should have an accurate record of state at that point in time.

What we need to do now, to keep later events from incorrectly being removed after compression is to update the event time during the compression, which will do in the next commit.